diff: Keep user's background color
[why]
The background in the main area is hardcoded to dark grey (30,30,30).
This breaks the usability for people who have a light theme (i.e. black
or dark text on light background).

[how]
One solution would be to also change the foreground color, but that will
for sure irritate people who use 'solarized' or other themes.

Imho it would be best not to tinker with the background colors unless at
least light and dark is selectable as option (and ideally autoselect
depending on the used terminal's current background (vim for example
does that, adapt the color scheme to the current user selected
background in the terminal).

So the solution implemented here is to completely drop the change in
background color.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>